Types: Update types definition
